src/main/java/com/medium/junewookang/springboot/web/HelloController.java
package com.medium.junewookang.springboot.web;

import com.medium.junewookang.springboot.web.dto.HelloResponseDto;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

@RestController
public class HelloController {
    @GetMapping("/hello")
    public String hello(){
        return "hello";
    }

    @GetMapping("/hello/dto")
    public HelloResponseDto helloDto(@RequestParam("name") String name, @RequestParam("amount") int amount){
        return new HelloResponseDto(name, amount); // 요청을 받으면 그걸로 새 객체를 만드는거구나.
    }
}
